requirements.txt
discord.py
pytz
opencv-python
numpy
Pillow
